Create button activation, emoji and color collection view, habit creation

+++ HabitTracker/UI/Tracker/ViewModels/RecordVM.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 struct RecordVM {
-    let trackerID: Int16
+    let trackerID: Int
     let date: Date
 }
